refactor(nvim): migrate to LazyVim
