+++ nvim/lazyvim.json
@@ -1,0 +1,9 @@
+{
+  "extras": [
+
+  ],
+  "news": {
+    "NEWS.md": "7107"
+  },
+  "version": 6
+}
