--- nvim/lua/binhtran432k/plugins/conform.lua
@@ -1,87 +1,0 @@
-local prettier = "prettierd"
-
-local function smart_fmt(bufnr)
-  local has_biome_lsp = vim.lsp.get_clients({
-    bufnr = bufnr,
-    name = "biome",
-  })[1]
-  if has_biome_lsp then
-    return { "biome" }
-  end
-  local has_denols = vim.lsp.get_clients({
-    bufnr = bufnr,
-    name = "denols",
-  })[1]
-  if has_denols then
-    return {}
-  end
-  local has_prettier = vim.fs.find({
-    -- https://prettier.io/docs/en/configuration.html
-    ".prettierrc",
-    ".prettierrc.json",
-    ".prettierrc.yml",
-    ".prettierrc.yaml",
-    ".prettierrc.json5",
-    ".prettierrc.js",
-    ".prettierrc.cjs",
-    ".prettierrc.toml",
-    "prettier.config.js",
-    "prettier.config.cjs",
-  }, { upward = true })[1]
-  if has_prettier then
-    return { prettier }
-  end
-  return { "biome" }
-end
-
-return { -- Autoformat
-  "stevearc/conform.nvim",
-  event = { "BufWritePre" },
-  cmd = { "ConformInfo" },
-  keys = {
-    {
-      "<leader>cf",
-      function()
-        require("conform").format({ async = true, lsp_format = "fallback" })
-      end,
-      mode = { "n", "v", "x" },
-      desc = "Code [F]ormat",
-    },
-  },
-  opts = {
-    notify_on_error = false,
-    -- format_on_save = function(bufnr)
-    --   -- Disable "format_on_save lsp_fallback" for languages that don't
-    --   -- have a well standardized coding style. You can add additional
-    --   -- languages here or re-enable it for the disabled ones.
-    --   local disable_filetypes = { c = true, cpp = true, zig = true }
-    --   local lsp_format_opt
-    --   if disable_filetypes[vim.bo[bufnr].filetype] then
-    --     lsp_format_opt = "never"
-    --   else
-    --     lsp_format_opt = "fallback"
-    --   end
-    --   return {
-    --     timeout_ms = 500,
-    --     lsp_format = lsp_format_opt,
-    --   }
-    -- end,
-    formatters_by_ft = {
-      css = smart_fmt,
-      graphql = smart_fmt,
-      html = { prettier },
-      javascript = smart_fmt,
-      javascriptreact = smart_fmt,
-      json = smart_fmt,
-      jsonc = smart_fmt,
-      lua = { "stylua" },
-      markdown = { prettier },
-      nix = { "alejandra" },
-      typescript = smart_fmt,
-      typescriptreact = smart_fmt,
-      yaml = { prettier },
-      -- Conform can also run multiple formatters sequentially
-      -- python = { "isort", "black" },
-    },
-  },
-}
