--- nvim/lua/binhtran432k/plugins/dressing.lua
@@ -1,5 +1,0 @@
-return {
-  "stevearc/dressing.nvim",
-  event = "VeryLazy",
-  opts = {},
-}
