--- nvim/lua/binhtran432k/plugins/guess_indent.lua
@@ -1,5 +1,0 @@
-return {
-  "nmac427/guess-indent.nvim",
-  event = { "BufReadPre", "BufNewFile" },
-  opts = {},
-}
