--- nvim/lua/binhtran432k/plugins/telescope.lua
@@ -1,128 +1,0 @@
-return {
-  {
-    -- Fuzzy Finder (files, lsp, etc)
-    "nvim-telescope/telescope.nvim",
-    branch = "0.1.x",
-    cmd = { "Telescope" },
-    dependencies = {
-      "nvim-lua/plenary.nvim",
-      "nvim-telescope/telescope-fzf-native.nvim",
-      "mollerhoj/telescope-recent-files.nvim",
-      -- Useful for getting pretty icons, but requires a Nerd Font.
-      { "nvim-tree/nvim-web-devicons", enabled = vim.g.have_nerd_font },
-    },
-    keys = {
-      { "<leader>sh", "<cmd>Telescope help_tags<cr>", desc = "[S]earch [H]elp" },
-      { "<leader>sk", "<cmd>Telescope keymaps<cr>", desc = "[S]earch [K]eymaps" },
-      { "<leader>sf", "<cmd>Telescope find_files<cr>", desc = "[S]earch [F]iles" },
-      { "<leader>ss", "<cmd>Telescope builtin<cr>", desc = "[S]earch [S]elect Telescope" },
-      { "<leader>sw", "<cmd>Telescope grep_string<cr>", desc = "[S]earch current [W]ord" },
-      { "<leader>sg", "<cmd>Telescope live_grep<cr>", desc = "[S]earch by [G]rep" },
-      { "<leader>sd", "<cmd>Telescope diagnostics<cr>", desc = "[S]earch [D]iagnostics" },
-      { "<leader>sr", "<cmd>Telescope resume<cr>", desc = "[S]earch [R]esume" },
-      { "<leader>s.", "<cmd>Telescope oldfiles<cr>", desc = '[S]earch Recent Files ("." for repeat)' },
-      { "<leader><leader>", "<cmd>Telescope buffers<cr>", desc = "[ ] Find existing buffers" },
-      -- Slightly advanced example of overriding default behavior and theme
-      {
-        "<leader>/",
-        function()
-          -- You can pass additional configuration to Telescope to change the theme, layout, etc.
-          require("telescope.builtin").current_buffer_fuzzy_find(require("telescope.themes").get_dropdown({
-            winblend = 10,
-            previewer = false,
-          }))
-        end,
-        desc = "[/] Fuzzily search in current buffer",
-      },
-      -- It's also possible to pass additional configuration options.
-      --  See `:help telescope.builtin.live_grep()` for information about particular keys
-      {
-        "<leader>s/",
-        function()
-          require("telescope.builtin").live_grep({
-            grep_open_files = true,
-            prompt_title = "Live Grep in Open Files",
-          })
-        end,
-        desc = "[S]earch [/] in Open Files",
-      },
-      -- Shortcut for searching your Neovim configuration files
-      {
-        "<leader>sn",
-        function()
-          require("telescope.builtin").find_files({ cwd = vim.fn.stdpath("config") })
-        end,
-        desc = "[S]earch [N]eovim files",
-      },
-    },
-    config = function()
-      -- Telescope is a fuzzy finder that comes with a lot of different things that
-      -- it can fuzzy find! It's more than just a "file finder", it can search
-      -- many different aspects of Neovim, your workspace, LSP, and more!
-      --
-      -- The easiest way to use Telescope, is to start by doing something like:
-      --  :Telescope help_tags
-      --
-      -- After running this command, a window will open up and you're able to
-      -- type in the prompt window. You'll see a list of `help_tags` options and
-      -- a corresponding preview of the help.
-      --
-      -- Two important keymaps to use while in Telescope are:
-      --  - Insert mode: <c-/>
-      --  - Normal mode: ?
-      --
-      -- This opens a window that shows you all of the keymaps for the current
-      -- Telescope picker. This is really useful to discover what Telescope can
-      -- do as well as how to actually do it!
-
-      -- [[ Configure Telescope ]]
-      -- See `:help telescope` and `:help telescope.setup()`
-      require("telescope").setup({
-        -- You can put your default mappings / updates / etc. in here
-        --  All the info you're looking for is in `:help telescope.setup()`
-        --
-        -- defaults = {
-        --   mappings = {
-        --     i = { ['<c-enter>'] = 'to_fuzzy_refine' },
-        --   },
-        -- },
-        -- pickers = {}
-        extensions = {
-          ["ui-select"] = {
-            require("telescope.themes").get_dropdown(),
-          },
-          frecency = { db_safe_mode = false },
-        },
-      })
-    end,
-  },
-  { -- If encountering errors, see telescope-fzf-native README for installation instructions
-    "nvim-telescope/telescope-fzf-native.nvim",
-    -- `build` is used to run some command when the plugin is installed/updated.
-    -- This is only run then, not every time Neovim starts up.
-    build = "make",
-    -- `cond` is a condition used to determine whether this plugin should be
-    -- installed and loaded.
-    cond = function()
-      return vim.fn.executable("make") == 1
-    end,
-    config = function()
-      require("telescope").load_extension("fzf")
-    end,
-  },
-  {
-    "mollerhoj/telescope-recent-files.nvim",
-    keys = {
-      {
-        "<leader>f",
-        function()
-          require("telescope").extensions["recent-files"].recent_files({})
-        end,
-        desc = "Open [F]recency",
-      },
-    },
-    config = function()
-      require("telescope").load_extension("recent-files")
-    end,
-  },
-}
