--- nvim/lua/binhtran432k/plugins/todo_comments.lua
@@ -1,7 +1,0 @@
--- Highlight todo, notes, etc in comments
-return {
-  "folke/todo-comments.nvim",
-  event = "VeryLazy",
-  dependencies = { "nvim-lua/plenary.nvim" },
-  opts = { signs = false },
-}
