+++ nvim/lua/plugins/core.lua
@@ -1,0 +1,4 @@
+return {
+  { "folke/lazy.nvim", version = false },
+  { "LazyVim/LazyVim", version = false },
+}
