+++ nvim/lua/plugins/util.lua
@@ -1,0 +1,3 @@
+return {
+  { import = "lazyvim.plugins.extras.util.mini-hipatterns" },
+}
