+++ nvim/spell/en.utf-8.add
@@ -1,0 +1,9 @@
+github
+configs
+LSP
+autogenerated
+neovim
+nvim
+lua
+binhtran432k
+io
